[fix] 'Resume' functionality does not resume in expected time if computer falls to sleep.
Explanation: When a Ubuntu computer goes to sleep, the "pause" time also pauses.
For example: Pause 5 min -> immediately go to sleep for any period -> wake-up = the connection paused and will be resumed in 5 mins.
This appears to be a bug in the 'setTimer()' function from Node.js. It seems to operate differently on various systems. This issue is not present on Windows.