(Linux) Build script improvement: ability to disable WIFI functionality (skipping "libiw-dev")
Currently, Wireless Tools (package "libiw-dev") needs to be available on the build machine.
We need it for the WIFI functionality being included in the daemon.

Some users may face issues building daemon due to missing "libiw-dev" on their systems.

So, now there is the possibility to skip using "libiw-dev" and disable WIFI functionality from the daemon build.
Environment variable `IVPN_NO_WIFI` must be defined for that.

Example:
```
  cd desktop-app/daemon/References/Linux/scripts/
  IVPN_NO_WIFI=true ./build-daemon.sh
```

https://github.com/ivpn/desktop-app/issues/345